Fix a bug when a directory element branches out
Directory elements are handled by looking at the differences between successive
versions, so that renames and deletes are seen.

But there is no link between the creation of a clearcase instance of a branch
and the global creation of the corresponding git branch. So we must be very
careful that the "previous version" of the directory is the one that was
actually current when the branch was created, including the cases where the
directory didn't exist at all at that time.