Handle cycles in branches
In clearcase, all the version trees are independent between elements. It is
therefore possible for an element to have branch branch_2 spawned from
branch_1, but the other way around for another element. It is of course not
possible to replicate this in git, so we chose one option and hope for the
best (which will likely mean some commits somewhat different from what
the "standard" clearcase view would have shown when the elements were
actually checked in.